+++ SRC/FUN/CC/Ensamblador.py
@@ -2,6 +2,7 @@ from FUN.CC.Ensamblador_1_directivas import *
 from FUN.CC.Ensamblador_2_segmento_datos import *
 from FUN.CC.Ensamblador_3_etiquetas import *
 from FUN.CC.Ensamblador_4_segmento_codigo import *
+from FUN.CC.Assembler25_stack import *
 from FUN.CC.Listado import *
 import os
 
@@ -16,15 +17,28 @@ def verificacion_codigo(DATOS: list, name: str):
     m_prog = {}
     listado = {}
     tabla_simbolos = []
+    simbolos = []
+    valores = {}
+    orig = {   "d": None,
+                                "s": None,
+                                "c": None}
 
     errores, mensaje, origen = verificar_directivas(DATOS)
 
     if errores == 0:
-        errores, mensaje_dseg, tabla_simbolos, listado_sim, cont_prog = verificar_segmento_datos(DATOS, origen)
+        errores, mensaje_dseg, tabla_simbolos, listado_sim,cont_prog, simbolos, valores, orig["d"] = verificar_segmento_datos(DATOS, origen)
     if errores == 0:
         tabla_simbolos, listado_etq = verificar_etiquetas(DATOS, origen, tabla_simbolos, cont_prog)
-        errores, mensaje_cseg, m_prog, listado_prog = verificar_segmento_codigo(DATOS, origen, tabla_simbolos, cont_prog)
-        mensaje = mensaje + mensaje_cseg + mensaje_dseg
+        orig["s"] = cont_prog//16
+        errores, msg_sseg, cont_prog, to_cseg, st_dir = validate_sseg(DATOS, origen,
+                tabla_simbolos, simbolos, valores, cont_prog)
+        if not to_cseg:
+            orig["s"] = None
+    if errores == 0:
+        orig["c"] = cont_prog//16
+        errores, mensaje_cseg, m_prog, listado_prog = verificar_segmento_codigo(DATOS, origen, tabla_simbolos, cont_prog,
+                to_cseg, st_dir)
+        mensaje = mensaje + mensaje_cseg + msg_sseg + mensaje_dseg
     if errores == 0:
         # Se agrega la tabla de símbolos a la memoria
         for i in range(len(tabla_simbolos)):
@@ -64,7 +78,7 @@ def verificacion_codigo(DATOS: list, name: str):
             listado_prog[i][1] = ' '.join(listado_prog[i][1])
         listado |= listado_prog
 
-    return errores, mensaje, m_prog, listado, tabla_simbolos, lib_info
+    return errores, mensaje, m_prog, listado, tabla_simbolos, lib_info, orig
 
 def expand_data(DATOS: list, data_fin: list, dir: str, lib_info : list):
     for i in range(len(DATOS)):
@@ -81,7 +95,7 @@ def expand_data(DATOS: list, data_fin: list, dir: str, lib_info : list):
             data_fin.append(DATOS[i])
         elif len(DATOS[i]) == 2 and DATOS[i][0].upper() == ".LIB":
             lib_fin = []
-            with open(os.path.join(os.path.dirname(dir), DATOS[i][1])) as archivo:
+            with open(os.path.join(os.path.dirname(dir), DATOS[i][1]), encoding="utf-8") as archivo:
                 lib_data = archivo.readlines()
                 lib_info.append([i, DATOS[i][1], os.path.join(os.path.dirname(dir), DATOS[i][1])])
             expand_data(lib_data, lib_fin, dir, lib_info)
